refactor(baseball): 주요기능이 있는 core 패키지 생성 후 이동
