+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.core.Baseball;
+
 public class Application {
 	public static void main(String[] args) {
 		// TODO 숫자 야구 게임 구현
